Flask_Crud/requirements.txt
attrs==24.2.0
blinker==1.8.2
boto3==1.35.54
botocore==1.35.54
click==8.1.7
colorama==0.4.6
Flask==3.0.3
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
MarkupSafe==3.0.2
mistune==3.0.2
packaging==24.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
referencing==0.35.1
rpds-py==0.20.1
s3transfer==0.10.3
six==1.16.0
urllib3==2.2.3
Werkzeug==3.1.2
pydantic>=2.0.0
email-validator>=2.0.0
python-jose[cryptography]
passlib
python-jwt
